Transcendance01/backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
django-cors-headers
